move verify fn to utils
